avr: Build configuration options for AVRSim.
	- Build can easily be configured using `ccmake`;
	- Options allow to use stdout or IPC queues to print status
	  messages;
	- `ipc.c` added to keep functions sending IPC messages together;
	- Several compiler warnings removed.

+++ avrsim/ihex.c
@@ -96,7 +96,7 @@ int Read_IHexRecord(IHexRecord *ihexRecord, FILE *in) {
 	/* Copy the ASCII hex encoding of the count field into hexBuff, convert it to a usable integer */
 	strncpy(hexBuff, recordBuff+IHEX_COUNT_OFFSET, IHEX_COUNT_LEN);
 	hexBuff[IHEX_COUNT_LEN] = 0;
-	dataCount = strtol(hexBuff, (char **)NULL, 16);
+	dataCount = (int) strtol(hexBuff, (char **)NULL, 16);
 
 	/* Copy the ASCII hex encoding of the address field into hexBuff, convert it to a usable integer */
 	strncpy(hexBuff, recordBuff+IHEX_ADDRESS_OFFSET, IHEX_ADDRESS_LEN);
@@ -106,7 +106,7 @@ int Read_IHexRecord(IHexRecord *ihexRecord, FILE *in) {
 	/* Copy the ASCII hex encoding of the address field into hexBuff, convert it to a usable integer */
 	strncpy(hexBuff, recordBuff+IHEX_TYPE_OFFSET, IHEX_TYPE_LEN);
 	hexBuff[IHEX_TYPE_LEN] = 0;
-	ihexRecord->type = strtol(hexBuff, (char **)NULL, 16);
+	ihexRecord->type = (int) strtol(hexBuff, (char **)NULL, 16);
 
 	/* Size check for start code, count, address, type, data and checksum fields */
 	if (strlen(recordBuff) < (unsigned int)(1+IHEX_COUNT_LEN+IHEX_ADDRESS_LEN+IHEX_TYPE_LEN+dataCount*2+IHEX_CHECKSUM_LEN))
@@ -184,8 +184,8 @@ uint8_t Checksum_IHexRecord(const IHexRecord *ihexRecord) {
 	int i;
 
 	/* Add the data count, type, address, and data bytes together */
-	checksum = ihexRecord->dataLen;
-	checksum += ihexRecord->type;
+	checksum = (uint8_t) ihexRecord->dataLen;
+	checksum += (uint8_t) ihexRecord->type;
 	checksum += (uint8_t)ihexRecord->address;
 	checksum += (uint8_t)((ihexRecord->address & 0xFF00)>>8);
 	for (i = 0; i < ihexRecord->dataLen; i++)
